requirements.txt
aioredis==1.3.0
appdirs==1.4.3
APScheduler==3.6.3
asgi-redis==1.4.3
asgiref==3.2.3
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.3.0
autobahn==19.11.1
Automat==0.8.0
azure-cognitiveservices-vision-face==0.4.1
azure-common==1.1.25
backcall==0.1.0
beautifulsoup4==4.8.2
bleach==3.1.0
bs4==0.0.1
CacheControl==0.12.6
cachetools==4.1.1
certifi==2019.11.28
cffi==1.13.2
channels==2.3.1
channels-redis==2.4.1
chardet==3.0.4
Click==7.0
colorama==0.4.1
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
cycler==0.10.0
daphne==2.4.0
decorator==4.4.2
defusedxml==0.5.0
distlib==0.3.0
Django==2.2.11
django-allauth==0.41.0
django-cors-headers==3.2.1
django-model-utils==4.0.0
django-multiselectfield==0.1.12
django-rest-auth==0.9.5
djangorestframework==3.11.0
drf-yasg==1.17.1
EasyProcess==0.2.10
entrypoints==0.3
fake-useragent==0.1.11
filelock==3.0.12
firebase-admin==4.3.0
flex==6.14.1
future==0.18.2
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth==1.18.0
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-firestore==1.7.0
google-cloud-storage==1.29.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
grpcio==1.30.0
hiredis==1.0.1
httplib2==0.18.1
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.6.0
incremental==17.5.0
inflection==0.3.1
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isodate==0.6.0
itypes==1.1.0
jedi==0.13.2
Jinja2==2.10
jsonpointer==2.0
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.2.0
lxml==4.4.2
MarkupSafe==1.1.0
matplotlib==3.2.1
mistune==0.8.4
more-itertools==8.2.0
msgpack==0.6.2
msgpack-python==0.5.6
msrest==0.6.19
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.4
numpy==1.18.4
oauthlib==3.1.0
opencv-python==4.2.0.34
packaging==20.1
pandocfilters==1.4.2
parso==0.3.1
pdfminer.six==20200402
pickleshare==0.7.5
pluggy==0.13.1
prometheus-client==0.5.0
prompt-toolkit==2.0.7
protobuf==3.12.2
psycopg2==2.8.4
psycopg2-binary==2.8.4
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
pycryptodome==3.9.7
Pygments==2.3.1
PyHamcrest==1.9.0
pyOpenSSL==19.1.0
pyparsing==2.4.6
pytest==5.4.1
python-dateutil==2.7.5
python-docx==0.8.10
python-telegram-bot==12.4.2
python3-openid==3.1.0
pytz==2019.3
PyVirtualDisplay==0.2.5
pywinpty==0.5.5
PyYAML==5.3
pyzmq==17.1.2
qtconsole==4.4.3
redis==2.10.6
requests==2.23.0
requests-oauthlib==1.3.0
rfc3987==1.3.8
rsa==4.6
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
six==1.12.0
sortedcontainers==2.1.0
soupsieve==2.0
sqlparse==0.3.1
strict-rfc3339==0.7
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
Twisted==19.10.0
txaio==18.8.1
tzlocal==2.0.0
ua-parser==0.10.0
uritemplate==3.0.1
urllib3==1.25.8
user-agents==2.1
validate-email==1.3
virtualenv==20.0.18
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
xvfbwrapper==0.2.9
zipp==3.1.0
zope.interface==4.7.1
